+++ client-hc/src/test/java/com/graphhopper/api/Examples.java
@@ -38,7 +38,7 @@ public class Examples {
                 addPoint(new GHPoint(49.6550, 11.4180));
         // Set profile like car, bike, foot, ...
         req.setProfile("bike");
-        // Optionally enable/disable elevation in output PointList, currently bike and foot support elevation, default is false
+        // Optionally enable/disable elevation in output PointList, default is false
         req.putHint("elevation", false);
         // Optionally enable/disable turn instruction information, defaults is true
         req.putHint("instructions", true);