+++ core/src/main/java/com/graphhopper/search/EdgeKVStorage.java
@@ -82,14 +82,10 @@ public class EdgeKVStorage {
     private long lastEntryPointer = -1;
     private List<KeyValue> lastEntries;
 
-    public EdgeKVStorage(Directory dir) {
-        this(dir, 1000);
-    }
-
     /**
      * Specify a larger cacheSize to reduce disk usage. Note that this increases the memory usage of this object.
      */
-    public EdgeKVStorage(Directory dir, final int cacheSize) {
+    public EdgeKVStorage(Directory dir) {
         keys = dir.create("edgekv_keys", 10 * 1024);
         vals = dir.create("edgekv_vals");
     }
@@ -177,6 +173,8 @@ public class EdgeKVStorage {
             if (key == null) throw new IllegalArgumentException("key cannot be null");
             Object value = entry.value;
             if (value == null) throw new IllegalArgumentException("value for key " + key + " cannot be null");
+            if (!entry.fwd && !entry.bwd)
+                throw new IllegalArgumentException("Do not add KeyValue pair where fwd and bwd is false");
             Integer keyIndex = keyToIndex.get(key);
             Class<?> clazz;
             if (keyIndex == null) {
@@ -475,5 +473,14 @@ public class EdgeKVStorage {
         public String toString() {
             return key + '=' + value + " (" + fwd + "|" + bwd + ")";
         }
+    }
+
+    /**
+     * This method limits the specified String value to the length currently accepted for values in the EdgeKVStorage.
+     */
+    public static String cutString(String value) {
+        byte[] bytes = value.getBytes(Helper.UTF_CS);
+        // See #2609 and test why we use a value < 255
+        return bytes.length > 250 ? new String(bytes, 0, 250, Helper.UTF_CS) : value;
     }
 }