+++ core/src/main/java/com/graphhopper/util/details/EdgeIdDetails.java
@@ -18,6 +18,7 @@
 package com.graphhopper.util.details;
 
 import com.graphhopper.routing.querygraph.VirtualEdgeIteratorState;
+import com.graphhopper.util.EdgeIterator;
 import com.graphhopper.util.EdgeIteratorState;
 import com.graphhopper.util.GHUtility;
 
@@ -30,7 +31,7 @@ import static com.graphhopper.util.Parameters.Details.EDGE_ID;
  */
 public class EdgeIdDetails extends AbstractPathDetailsBuilder {
 
-    private int edgeId = -1;
+    private int edgeId = EdgeIterator.NO_EDGE;
 
     public EdgeIdDetails() {
         super(EDGE_ID);